% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.contigs.R
\name{select.contigs}
\alias{select.contigs}
\title{Filter and sample contigs based on SNP density and missing data}
\usage{
select.contigs(
  wd = getwd(),
  kept.sites,
  n.loci,
  snpden,
  SNP.density.range = c(0.05, 0.95),
  lmiss,
  min.lmiss,
  out.name = paste(kept.sites, ".out", sep = "")
)
}
\arguments{
\item{wd}{Character. Working directory. Default is \code{getwd()}.}

\item{kept.sites}{Character. Path to a tab-delimited file containing retained sites (in vcftools format).}

\item{n.loci}{Integer or \code{"all"}. Number of loci to retain after filtering.}

\item{snpden}{Character or NULL. Path to a file with SNP counts per contig (in vcftools format).
If \code{NULL}, SNP density filtering is skipped.}

\item{SNP.density.range}{Numeric vector of length 2. Quantile range used to
filter contigs based on SNP density (default = \code{c(0.05, 0.95)}).}

\item{lmiss}{Character or NULL. Path to file with missing data per SNP (in vcftools format).
If \code{NULL}, missing data filtering is skipped.}

\item{min.lmiss}{Numeric. Maximum allowed missing data per SNP.}

\item{out.name}{Character. Output file name.}
}
\value{
No return value. Writes filtered loci to file.
}
\description{
This function filters contigs (loci) based on SNP density and/or missing data,
and optionally subsamples a fixed number of loci. The filtered dataset is
written to a tab-delimited file. Kept sites, SNP density and missing data input are in vcftools format (https://vcftools.github.io/man_latest.html)
}
\details{
The function applies the following filters:
\itemize{
\item Removes contigs outside the specified SNP density quantile range.
\item Removes contigs with SNPs exceeding the allowed missing data threshold.
\item Optionally subsamples a fixed number of loci.
}
}
\examples{
\dontrun{
select.contigs(
  kept.sites = "sites.txt",
  n.loci = 100,
  snpden = "snp_density.txt",
  lmiss = "missing.txt",
  min.lmiss = 0.2
)
}

}
