% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf2longLoci.R
\name{vcf2longLoci}
\alias{vcf2longLoci}
\title{Convert VCF data into long-format loci sequences}
\usage{
vcf2longLoci(
  wd = getwd,
  vcf.name,
  kept.sites,
  reference,
  rm.initial.seq = "",
  rm.final.seq = "",
  phased.vcf = T,
  out.name
)
}
\arguments{
\item{wd}{Character. Working directory. Default is \code{getwd}.}

\item{vcf.name}{Character. Path to the VCF file.}

\item{kept.sites}{Character. File containing retained SNP positions.}

\item{reference}{Character. Path to a reference FASTA file.}

\item{rm.initial.seq}{Character. Sequence to remove from the beginning
of each reconstructed sequence. For example, to delete EcoRI sequence "NTAATTC" (default = \code{""}).}

\item{rm.final.seq}{Character. Sequence to remove from the end of each
reconstructed sequence (default = \code{""}).}

\item{phased.vcf}{Logical. Whether the VCF is phased (default = \code{TRUE}).}

\item{out.name}{Character. Output file name.}
}
\value{
No return value. Writes reconstructed sequences to file.
}
\description{
This function reconstructs full-length sequences for each individual
and contig from a VCF file and a reference genome. It outputs sequences
in a long format suitable for downstream phylogenetic or population
genetic analyses. Kept sites is in vcftools format (https://vcftools.github.io/man_latest.html) and reference in dDocent fasta format (https://ddocent.com//)
}
\details{
The function:
\itemize{
\item Reads SNP positions and VCF genotype data.
\item Reconstructs sequences using the reference genome.
\item Handles phased and unphased data differently:
\itemize{
\item Phased: reconstructs two haplotypes per individual.
\item Unphased: encodes heterozygous sites using IUPAC ambiguity codes.
}
\item Replaces missing data ('.') with '?'.
\item Outputs sequences in a simple text format.
}
}
\examples{
\dontrun{
vcf2longLoci(
  vcf.name = "data.vcf",
  kept.sites = "sites.txt",
  reference = "ref.fasta",
  phased.vcf = TRUE,
  out.name = "output.txt"
)
}

}
